Write tensor as float32 instead of float64 to allow training with float32 weight Tensors.

+++ write_tensors.py
@@ -22,7 +22,7 @@ def print_shapes(c_data, n_data):
 
 
 def write_tensor(arr, filename):
-    tensor = torch.from_numpy(arr)
+    tensor = torch.from_numpy(arr).float()
     torch.save(tensor, filename)
